requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.0.13
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.11.0
gunicorn==19.7.1
idna==2.6
oauthlib==3.1.0
mysqlclient==1.4.6
python3-openid==3.1.0
pytz==2019.3
requests==2.18.4
requests-oauthlib==1.3.0
six==1.9.0
urllib3==1.22
whitenoise==4.0
psycopg2==2.8.5
dj-database-url==0.5.0

